Move Gatherer and ListGatherer to mutate.cpp
